fix KeyValueID type

+++ lib/src/main/kotlin/dev/klerkframework/klerk/storage/Persistence.kt
@@ -38,13 +38,13 @@ public interface Persistence {
     public fun setConfig(config: Config<*, *>): Unit
     public fun migrate(migrations: List<MigrationStep>): Unit
 
-    public fun putKeyValue(id: UInt, value: String, ttl: Instant?): Unit
-    public fun putKeyValue(id: UInt, value: Int, ttl: Instant?): Unit
-    public fun putKeyValue(id: UInt, value: InputStream, ttl: Instant?): Unit
-    public fun updateBlob(id: UInt, ttl: Instant?, active: Boolean): Unit
-    public fun getKeyValueString(id: UInt): Pair<String, Instant?>?
-    public fun getKeyValueInt(id: UInt): Pair<Int, Instant?>?
-    public fun getKeyValueBlob(id: UInt): Triple<InputStream, Instant?, Boolean>?
+    public fun putKeyValue(id: Long, value: String, ttl: Instant?): Unit
+    public fun putKeyValue(id: Long, value: Int, ttl: Instant?): Unit
+    public fun putKeyValue(id: Long, value: InputStream, ttl: Instant?): Unit
+    public fun updateBlob(id: Long, ttl: Instant?, active: Boolean): Unit
+    public fun getKeyValueString(id: Long): Pair<String, Instant?>?
+    public fun getKeyValueInt(id: Long): Pair<Int, Instant?>?
+    public fun getKeyValueBlob(id: Long): Triple<InputStream, Instant?, Boolean>?
 }
 
 /**
@@ -54,9 +54,9 @@ public class RamStorage : Persistence {
     private val auditLog = mutableSetOf<AuditEntry>()
     private val models = mutableMapOf<Int, Model<Any>>()
     override val currentModelSchemaVersion: Int = 1
-    private val keyValueStrings = mutableMapOf<UInt, Pair<String, Instant?>>()
-    private val keyValueInts = mutableMapOf<UInt, Pair<Int, Instant?>>()
-    private val keyValueBlobs = mutableMapOf<UInt, Triple<ByteArray, Instant?, Boolean>>()
+    private val keyValueStrings = mutableMapOf<Long, Pair<String, Instant?>>()
+    private val keyValueInts = mutableMapOf<Long, Pair<Int, Instant?>>()
+    private val keyValueBlobs = mutableMapOf<Long, Triple<ByteArray, Instant?, Boolean>>()
 
     override fun <T : Any, P, C:KlerkContext, V> store(
         delta: ProcessingData<out T, C, V>,
@@ -115,19 +115,19 @@ public class RamStorage : Persistence {
         logger.debug { "Skipping migration since RamStorage is always empty on startup" }
     }
 
-    override fun putKeyValue(id: UInt, value: String, ttl: Instant?) {
+    override fun putKeyValue(id: Long, value: String, ttl: Instant?) {
         keyValueStrings[id] = Pair(value, ttl)
     }
 
-    override fun putKeyValue(id: UInt, value: Int, ttl: Instant?) {
+    override fun putKeyValue(id: Long, value: Int, ttl: Instant?) {
         keyValueInts[id] = Pair(value, ttl)
     }
 
-    override fun putKeyValue(id: UInt, value: InputStream, ttl: Instant?) {
+    override fun putKeyValue(id: Long, value: InputStream, ttl: Instant?) {
         keyValueBlobs[id] = Triple(value.readAllBytes(), ttl, false)
     }
 
-    override fun updateBlob(id: UInt, ttl: Instant?, active: Boolean) {
+    override fun updateBlob(id: Long, ttl: Instant?, active: Boolean) {
         val old = keyValueBlobs[id]
         if (old == null) {
             logger.warn { "Could not find blob with id $id" }
@@ -136,9 +136,9 @@ public class RamStorage : Persistence {
         keyValueBlobs[id] = Triple(old.first, ttl, active)
     }
 
-    override fun getKeyValueString(id: UInt): Pair<String, Instant?>? = keyValueStrings[id]
-    override fun getKeyValueInt(id: UInt): Pair<Int, Instant?>? = keyValueInts[id]
-    override fun getKeyValueBlob(id: UInt): Triple<InputStream, Instant?, Boolean>? =
+    override fun getKeyValueString(id: Long): Pair<String, Instant?>? = keyValueStrings[id]
+    override fun getKeyValueInt(id: Long): Pair<Int, Instant?>? = keyValueInts[id]
+    override fun getKeyValueBlob(id: Long): Triple<InputStream, Instant?, Boolean>? =
         keyValueBlobs[id]?.let { Triple(it.first.inputStream(), it.second, it.third) }
 
     public fun <T : Any, P, C:KlerkContext, V> createAuditEntry(
